Refactoring of project structure
Basic stuff goes in Core, everything function-related goes in Function
