--- Source/CoreTest/src/ca/uqac/lif/petitpoucet/function/number/package-info.java
@@ -1,1 +1,0 @@
-package ca.uqac.lif.petitpoucet.function.number;
