+++ Source/Functions/src/ca/uqac/lif/petitpoucet/function/AtomicFunction.java
