+++ Source/Functions/src/ca/uqac/lif/petitpoucet/function/Constant.java
