+++ Source/Functions/src/ca/uqac/lif/petitpoucet/function/ExplanationQueryable.java
