+++ Source/Functions/src/ca/uqac/lif/petitpoucet/function/Fork.java
