+++ Source/Functions/src/ca/uqac/lif/petitpoucet/function/FunctionException.java
