+++ Source/Functions/src/ca/uqac/lif/petitpoucet/function/FunctionPin.java
