+++ Source/Functions/src/ca/uqac/lif/petitpoucet/function/InvalidNumberOfArgumentsException.java
