+++ Source/Functions/src/ca/uqac/lif/petitpoucet/function/number/Addition.java
