+++ Source/Functions/src/ca/uqac/lif/petitpoucet/function/number/Multiplication.java
