+++ Source/Functions/src/ca/uqac/lif/petitpoucet/function/number/package-info.java
