+++ Source/Functions/src/ca/uqac/lif/petitpoucet/function/package-info.java
