+++ Source/Functions/src/ca/uqac/lif/petitpoucet/function/vector/ParameterizedVectorFunction.java
