+++ Source/Functions/src/ca/uqac/lif/petitpoucet/function/vector/VectorSum.java
