+++ Source/Functions/src/ca/uqac/lif/petitpoucet/function/vector/Window.java
