+++ Source/Functions/src/ca/uqac/lif/petitpoucet/function/vector/package-info.java
