+++ Source/Functions/srctest/ca/uqac/lif/petitpoucet/function/ConstantTest.java
