+++ Source/Functions/srctest/ca/uqac/lif/petitpoucet/function/vector/NthElementTest.java
