+++ Source/Functions/srctest/ca/uqac/lif/petitpoucet/function/vector/VectorApplyTest.java
