+++ Source/Functions/srctest/ca/uqac/lif/petitpoucet/function/vector/VectorTestUtilities.java
